+++ common/testutils/random/test_random_test.go
@@ -1,4 +1,4 @@
-package testutils
+package random
 
 import (
 	"github.com/stretchr/testify/assert"
